Rename module stylesheets
