+++ src/components/PageWrapper/PageWrapper.tsx
@@ -3,7 +3,7 @@ import {Footer} from './Footer'
 import {Header} from './Header'
 import {HeroTitle} from './HeroTitle'
 import {SiteLogo} from './SiteLogo'
-import styles from './page-wrapper.module.scss'
+import styles from './PageWrapper.module.scss'
 import {ColorSchemeContext} from '@contexts/ColorScheme'
 import {useSafeContext} from '@hooks/use-safe-context'
 import {combineClassNames} from '@utilities/combine-class-names'
